chore: rename target_kind module to target
